Create semseg and tagging problem_types
* Create packages for different computer vision problem_types
* Switch on problem_type in the run.py and call the appropriate run_tasks function
* Extract code that is common across problem types and put in the common package
