+++ src/rastervision/common/tasks/plot_curves.py
