+++ src/rastervision/semseg/options.py
@@ -1,5 +1,3 @@
-import json
-
 from .models.conv_logistic import CONV_LOGISTIC
 from .models.fcn_resnet import FCN_RESNET
 from .models.dual_fcn_resnet import DUAL_FCN_RESNET
@@ -7,41 +5,15 @@ from .models.fc_densenet import FC_DENSENET
 from .models.ensemble import CONCAT_ENSEMBLE, AVG_ENSEMBLE
 from .data.potsdam import POTSDAM, PotsdamDataset
 from .data.vaihingen import VAIHINGEN
+from rastervision.common.options import Options
 
 
-class RunOptions():
+class SemsegOptions(Options):
     """Represents the options used to control an experimental run."""
 
     def __init__(self, options):
-        if 'train_stages' in options and options['train_stages'] is not None:
-            train_stages = options['train_stages']
-            options.update(train_stages[0])
-        self.train_stages = options.get('train_stages')
+        super().__init__(options)
 
-        # Required options
-        self.model_type = options['model_type']
-        self.run_name = options['run_name']
-        self.dataset_name = options['dataset_name']
-        self.generator_name = options['generator_name']
-        self.batch_size = options['batch_size']
-        self.epochs = options['epochs']
-        self.steps_per_epoch = options['steps_per_epoch']
-        self.validation_steps = options['validation_steps']
-        self.active_input_inds = options['active_input_inds']
-
-        # Optional options
-        self.git_commit = options.get('git_commit')
-        # Size of the imgs used as input to the network
-        # [nb_rows, nb_cols]
-        self.target_size = options.get('target_size', (256, 256))
-        self.optimizer = options.get('optimizer', 'adam')
-        self.init_lr = options.get('init_lr', 1e-3)
-        self.patience = options.get('patience')
-        self.lr_schedule = options.get('lr_schedule')
-        self.train_ratio = options.get('train_ratio')
-        self.cross_validation = options.get('cross_validation')
-        self.delta_model_checkpoint = options.get(
-            'delta_model_checkpoint', None)
         self.nb_videos = options.get('nb_videos')
 
         # Controls how many samples to use in the final evaluation.
@@ -49,8 +21,6 @@ class RunOptions():
         # the code, since it will save time.
         self.nb_eval_samples = options.get('nb_eval_samples')
 
-        # Model type dependent options
-        # TODO have models check/parse options instead of doing it here
         if self.model_type == CONV_LOGISTIC:
             self.kernel_size = options['kernel_size']
         elif self.model_type == FC_DENSENET:
@@ -78,12 +48,3 @@ class RunOptions():
         # None means to use the size of the original image.
         self.eval_target_size = options.get(
             'eval_target_size', default_eval_target_size)
-
-
-def load_options(file_path):
-    options = None
-    with open(file_path) as options_file:
-        options = json.load(options_file)
-        options = RunOptions(options)
-
-    return options
