+++ src/rastervision/semseg/settings.py
@@ -1,0 +1,1 @@
+semseg = 'semseg'
