+++ src/rastervision/tagging/run.py
@@ -1,0 +1,4 @@
+def run_tasks(options_dict, tasks):
+    # TODO
+    print(options_dict)
+    print(tasks)
