+++ src/rastervision/tagging/settings.py
@@ -1,0 +1,1 @@
+tagging = 'tagging'
